+++ platform/support/support.boot/src/main/java/de/iip_ecosphere/platform/support/xml/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Simplified Jackson-like XML mapping.
+ */
+package de.iip_ecosphere.platform.support.xml;