+++ platform/support/support.boot/src/test/java/test/de/iip_ecosphere/platform/support/xml/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Tests the XML abstraction.
+ */
+package test.de.iip_ecosphere.platform.support.xml;